Merge 2.x to master

+++ app/src/ts/renderer/mishiro-home.ts
@@ -3,14 +3,14 @@ import TheTable from '../../vue/component/TheTable.vue'
 import InputText from '../../vue/component/InputText.vue'
 import ProgressBar from '../../vue/component/ProgressBar.vue'
 
-import { Event } from 'electron'
+// import { Event } from 'electron'
 import { Vue, Component } from 'vue-property-decorator'
-import { generateObjectId } from '../common/object-id'
+// import { generateObjectId } from '../common/object-id'
 // import { ProgressInfo } from 'mishiro-core'
 const fs = window.node.fs
 const path = window.node.path
 const getPath = window.preload.getPath
-const { shell, ipcRenderer, remote } = window.node.electron
+const { shell, remote } = window.node.electron
 const { downloadDir } = getPath
 
 @Component({
@@ -34,7 +34,7 @@ export default class extends Vue {
   notDownloadedOnly: boolean = false
   canDownloadRows: any[] = []
 
-  get totalPage () {
+  get totalPage (): number {
     const canDownload: any[] = this.canDownloadRows
     if (!canDownload.length) return 0
     return canDownload.length / this.recordPerPage === Math.floor(canDownload.length / this.recordPerPage) ? canDownload.length / this.recordPerPage - 1 : Math.floor(canDownload.length / this.recordPerPage)
@@ -45,31 +45,37 @@ export default class extends Vue {
   //   return this.data.filter(row => !this.isDisabled(row))
   // }
 
-  checkFile (data: any[]): Promise<any[]> {
-    return new Promise<any[]>((resolve, reject) => {
-      const id = generateObjectId()
-      ipcRenderer.once('checkFile', (_ev: Event, oid: string, notDownloaded: any[]) => {
-        if (oid === id) {
-          resolve(notDownloaded)
-        } else {
-          reject(new Error(`${id} !== ${oid}`))
-        }
-      })
-      ipcRenderer.send('checkFile', id, data)
-    })
+  checkFile (data: any[]): any[] {
+    // return new Promise<any[]>((resolve, reject) => {
+    // const id = generateObjectId()
+    // ipcRenderer.once('checkFile', (_ev: Event, oid: string, notDownloaded: any[]) => {
+    //   if (oid === id) {
+    //     resolve(notDownloaded)
+    //   } else {
+    //     reject(new Error(`${id} !== ${oid}`))
+    //   }
+    // })
+    // ipcRenderer.send('checkFile', id, data)
+    // })
+    return data.filter(row => !fs.existsSync(getPath.downloadDir(path.basename(row.name))))
   }
 
-  isDisabled (row: any) {
+  isDisabled (row: any): boolean {
     return fs.existsSync(downloadDir(path.basename(row.name)))
   }
 
-  opendir () {
+  opendir (): void {
     this.playSe(this.enterSe)
     const dir = downloadDir()
     if (!fs.existsSync(dir)) fs.mkdirsSync(dir)
-    process.platform === 'win32' ? shell.openExternal(dir) : shell.showItemInFolder(dir + '/.')
+    if (process.platform === 'win32') {
+      shell.openExternal(dir).catch(err => console.log(err))
+    } else {
+      shell.showItemInFolder(dir + '/.')
+    }
   }
-  query () {
+
+  query (): void {
     if (this.queryString === '') {
       this.page = 0
       this.data = []
@@ -78,26 +84,47 @@ export default class extends Vue {
     } else if (this.queryString === 'dev') {
       remote.getCurrentWindow().webContents.openDevTools()
     } else {
-      ipcRenderer.send('queryManifest', this.queryString)
+      const manifestDB = window.preload.getManifestDB()
+      if (!manifestDB) return
+      manifestDB.find<{ name: string, hash: string }>('manifests', ['name', 'hash'], { name: { $like: `%${this.queryString.trim()}%` } }).then(manifestArr => {
+        this.page = 0
+        this.data = manifestArr
+        if (!this.notDownloadedOnly) {
+          this.canDownloadRows = this.data
+        } else {
+          // this.checkFile(this.data).then((res) => {
+          //   this.canDownloadRows = res
+          // }).catch(err => console.log(err))
+          this.canDownloadRows = this.checkFile(this.data)
+        }
+      }).catch(err => {
+        this.event.$emit('alert', this.$t('home.errorTitle'), err.message)
+      })
+      // ipcRenderer.send('queryManifest', this.queryString)
     }
-    this.playSe(this.enterSe)
+    // this.playSe(this.enterSe)
   }
-  filterOnClick () {
+
+  filterOnClick (): void {
     this.notDownloadedOnly = !this.notDownloadedOnly
     if (!this.notDownloadedOnly) {
       this.canDownloadRows = this.data
       this.page = 0
     } else {
-      this.checkFile(this.data).then((res) => {
-        this.canDownloadRows = res
-        this.page = 0
-      }).catch(err => console.log(err))
+      // this.checkFile(this.data).then((res) => {
+      //   this.canDownloadRows = res
+      //   this.page = 0
+      // }).catch(err => console.log(err))
+      this.canDownloadRows = this.checkFile(this.data)
+      this.page = 0
     }
   }
-  tableChange (val: any[]) {
+
+  tableChange (val: any[]): void {
     this.selectedItem = val
   }
-  stopDownload () {
+
+  stopDownload (): void {
     this.playSe(this.cancelSe)
     this.downloadBtnDisable = false
     this.total = 0
@@ -105,7 +132,8 @@ export default class extends Vue {
     this.text = ''
     this.dler.stop(() => this.event.$emit('alert', this.$t('home.errorTitle'), this.$t('home.noTask')))
   }
-  async downloadSelectedItem () {
+
+  async downloadSelectedItem (): Promise<void> {
     this.playSe(this.enterSe)
     if (!navigator.onLine) {
       this.event.$emit('alert', this.$t('home.errorTitle'), this.$t('home.noNetwork'))
@@ -116,7 +144,7 @@ export default class extends Vue {
     if (tasks.length > 0) {
       this.downloadBtnDisable = true
 
-      let errorList = await this.dler.batchDownload(
+      const errorList = await this.dler.batchDownload(
         tasks,
         downloadDir(),
         (_row, filepath) => {
@@ -136,7 +164,7 @@ export default class extends Vue {
 
           this.current = 0
           this.text = ''
-          if (suffix !== '.acb') {
+          if (suffix !== '.acb' && suffix !== '.awb') {
             if (fs.existsSync(filepath)) {
               fs.removeSync(filepath)
               this.event.$emit('completeTask', name + suffix)
@@ -204,7 +232,7 @@ export default class extends Vue {
     }
   }
 
-  onMouseWheel (e: WheelEvent) {
+  onMouseWheel (e: WheelEvent): void {
     if (e.deltaY < 0) {
       this.previousPage()
     } else {
@@ -212,32 +240,32 @@ export default class extends Vue {
     }
   }
 
-  previousPage () {
+  previousPage (): void {
     this.page !== 0 ? this.page -= 1 : this.page = this.totalPage
   }
 
-  nextPage () {
+  nextPage (): void {
     this.page !== this.totalPage ? this.page += 1 : this.page = 0
   }
 
-  mounted () {
+  mounted (): void {
     this.$nextTick(() => {
       this.event.$on('enterKey', (block: string) => {
         if (block === 'home') {
           this.query()
         }
       })
-      ipcRenderer.on('queryManifest', (_event: Event, manifestArr: any[]) => {
-        this.page = 0
-        this.data = manifestArr
-        if (!this.notDownloadedOnly) {
-          this.canDownloadRows = this.data
-        } else {
-          this.checkFile(this.data).then((res) => {
-            this.canDownloadRows = res
-          }).catch(err => console.log(err))
-        }
-      })
+      // ipcRenderer.on('queryManifest', (_event: Event, manifestArr: any[]) => {
+      //   this.page = 0
+      //   this.data = manifestArr
+      //   if (!this.notDownloadedOnly) {
+      //     this.canDownloadRows = this.data
+      //   } else {
+      //     this.checkFile(this.data).then((res) => {
+      //       this.canDownloadRows = res
+      //     }).catch(err => console.log(err))
+      //   }
+      // })
     })
   }
 }
